+++ src/Food_Item.css
@@ -1,106 +1,87 @@
-/* General container styles */
-.list_container {
-  font-family: Arial, sans-serif;
+.table-container {
+  margin: 10px;
   display: flex;
-  position: fixed;
-  right: 0px;
-  top: 170px;
-  width: 1240px;
+  justify-content: center;
+  gap: 30px;
+  height: 500px;
 }
 
-.list_header{
-    display: flex;
-    position: fixed;
-    left: 290px;
-    top: 100px;
-}
-/* Table styles */
-table {
-  width: 100%;
+.client-table {
+  width: 80%;
+  overflow-y: auto;
   border-collapse: collapse;
-  margin: 20px 0;
-}
-
-thead th {
-  background-color: #f4f4f4;
-  font-weight: bold;
-  padding: 10px;
-  border-bottom: 2px solid #ddd;
-}
-
-tbody td {
-  padding: 10px;
   text-align: left;
-  border-bottom: 1px solid #ddd;
-}
-
-/* Row hover effect */
-tbody tr:hover {
+  font-family: Arial, sans-serif;
   background-color: #f9f9f9;
+  position: fixed;
+  right: 0;
+  top: 130px;
+  height: auto;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
 }
 
-/* Buttons */
-button {
-  padding: 8px 15px;
-  margin-right: 10px;
+.client-table th {
+  background-color: #4caf50;
+  color: white;
+  text-transform: uppercase;
+  padding: 12px;
+  position: sticky;
+  top: 0;
+  z-index: 1;
+}
+
+.client-table tr:nth-child(even) {
+  background-color: #f2f2f2;
+}
+
+.client-table tr:hover {
+  background-color: #ddd;
+}
+
+.client-table td {
+  font-size: 16px;
+  padding: 12px;
+}
+
+button.approve,
+button.reject {
+  padding: 8px 12px;
   border: none;
-  border-radius: 5px;
-  cursor: pointer;
+  border-radius: 4px;
   font-size: 14px;
+  cursor: pointer;
+  margin-right: 5px;
 }
 
-button:first-of-type {
-  background-color: #4caf50; /* Green color for Assign Nutritionist */
-  color: white;
+button.approve {
+  background-color: #4caf50;
+  color: #fff;
 }
 
-button:last-of-type {
-  background-color: #f44336; /* Red color for Delete */
-  color: white;
+button.reject {
+  background-color: #f44336;
+  color: #fff;
+}
+
+button.approve:hover {
+  background-color: #45a049;
+}
+
+button.reject:hover {
+  background-color: #d32f2f;
+}
+
+button {
+  margin-bottom: 10px;
+  padding: 10px 20px;
+  font-size: 14px;
+  background-color: #3f51b5;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
 }
 
 button:hover {
-  opacity: 0.9;
-}
-
-/* Responsive adjustments */
-@media (max-width: 768px) {
-  table,
-  thead,
-  tbody,
-  th,
-  td,
-  tr {
-    display: block;
-    width: 100%;
-  }
-
-  th {
-    display: none;
-  }
-
-  tr {
-    margin-bottom: 15px;
-  }
-
-  td {
-    padding: 10px;
-    text-align: right;
-    position: relative;
-  }
-
-  td:before {
-    content: attr(data-label);
-    position: absolute;
-    left: 0;
-    padding-left: 10px;
-    font-weight: bold;
-    text-align: left;
-  }
-}
-
-.search_bar{
-  position: fixed;
-    right: 0px;
-    top: 145px;
+  background-color: #303f9f;
 }
